feat(controller): use event delegation for binding events

+++ javascript-practice/src/js/controllers/product.controller.js
@@ -1,5 +1,5 @@
 import Toast from '../utils/toastify';
-import { querySelectorAll } from '../utils/dom';
+import { addEventListener, querySelector } from '../utils/dom';
 import MESSAGES from '../constants/messages';
 
 export default class ProductController {
@@ -13,6 +13,8 @@ export default class ProductController {
    */
   async init() {
     await this.displayProducts();
+
+    this.bindDeleteProductEvent();
   }
 
   /**
@@ -22,8 +24,6 @@ export default class ProductController {
     const products = await this.service.getAll();
 
     this.view.renderProducts(products);
-
-    this.bindDeleteProductEvent();
   }
 
   /**
@@ -35,25 +35,29 @@ export default class ProductController {
       DELETE_PRODUCT_SUCCESS_MSG,
       DELETE_PRODUCT_FAILED_MSG
     } = MESSAGES;
-    const btnDeleteElements = querySelectorAll('.btn-delete');
+    const productListElement = querySelector('.main-products-container');
 
-    btnDeleteElements.forEach(element => {
-      element.addEventListener('click', async (e) => {
-        const target = e.target;
-        const id = target.dataset.id;
+    addEventListener(productListElement, 'click', async (e) => {
+      const { target } = e;
 
-        if (!confirm(MESSAGES.DELETE_CONFIRMATION_MSG)) {
-          return;
-        }
-        const { isSuccess } = await this.service.deleteById(id);
+      if (!target.classList.contains('btn-delete')) {
+        return;
+      }
 
-        if (!isSuccess) {
-          return Toast.error(DELETE_PRODUCT_FAILED_MSG);
-        }
+      const id = target.dataset.id;
 
-        Toast.success(DELETE_PRODUCT_SUCCESS_MSG);
-        this.displayProducts();
-      });
+      if (!confirm(MESSAGES.DELETE_CONFIRMATION_MSG)) {
+        return;
+      }
+      const { isSuccess } = await this.service.deleteById(id);
+
+      if (!isSuccess) {
+        return Toast.error(DELETE_PRODUCT_FAILED_MSG);
+      }
+
+      Toast.success(DELETE_PRODUCT_SUCCESS_MSG);
+      this.displayProducts();
+
     });
   }
 }
